requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.12
